fixed ad67388 from https://gitee.com/gongjunsong/ark_js_runtime/pulls/2526 Fix bug in Array.sort interface
Fix bug in Array.sort interface

issue:https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I5TKT2

+++ ecmascript/base/array_helper.h
@@ -30,8 +30,8 @@ struct FlattenArgs {
 class ArrayHelper {
 public:
     static bool IsConcatSpreadable(JSThread *thread, const JSHandle<JSTaggedValue> &obj);
-    static int32_t SortCompare(JSThread *thread, const JSHandle<JSTaggedValue> &callbackfnHandle,
-                               const JSHandle<JSTaggedValue> &valueX, const JSHandle<JSTaggedValue> &valueY);
+    static double SortCompare(JSThread *thread, const JSHandle<JSTaggedValue> &callbackfnHandle,
+                              const JSHandle<JSTaggedValue> &valueX, const JSHandle<JSTaggedValue> &valueY);
     static int64_t GetLength(JSThread *thread, const JSHandle<JSTaggedValue> &thisHandle);
     static int64_t GetArrayLength(JSThread *thread, const JSHandle<JSTaggedValue> &thisHandle);
     static JSTaggedValue FlattenIntoArray(JSThread *thread, const JSHandle<JSObject> &newArrayHandle,
